Moved www directory and all it's contents to top level.
Merged www/common and common.
This is required in order to get the site working with a commercial webhost.
Commercial webhosts (in this case, greengeeks) do not allow you to modify httpd.conf
so using the VirtualHost directive to redirect to the www won't work.
The easiest way to handle this was moving the www directory to the top level of
the site. Other changes are related to this change.